Avoid unnecessary datapack read
We write the KeyValues handle before the flags so we don't have to read
the flags cell to get at the handle.

This is preferred as when no valid players are present in the callbacks,
we can just read the next cell to get the handle that needs to be freed.